// src/app/models/tour-package-view.model.ts
export class TourPackageViewModel {
  currencyCode: string;
  adultNum: string;
  baseCost: string;
  taxesPerAdult: string;
  insurancePerAdult: string;
  childrenNumber: string;
  childBaseCost: string;
  taxesPerChild: string;
  insurancePerChild: string;
  infantNumber: string;
  totalCostPerInfant: string;
  depositPaid: string;
  totalCostHoliday: number;
  lessDepositPaid: string;
  balanceToBePaid: number;
  balanceDueDate: Date;
  commisionAmount: string;
}